+++ R/bootstrap_model.R
@@ -146,7 +146,45 @@ bootstrap_model <- function(base_model,
         }
     }
 
-    parallelism <- match.arg(parallelism)
+    if (missing(parallelism)) {
+        if (!is.null(num_cores) && num_cores > 1) {
+            ## if just the num_cores argument is used, we'll
+            ## assume parallel::mclapply is desired
+            if (!requireNamespace("parallel", quietly = TRUE)) {
+                stop("setting `num_cores` greater than 1 without setting ",
+                     "`parallelism` uses `package:parallel`, ",
+                     "but it's not installed", call. = FALSE)
+            }
+            parallelism <- "parallel"
+        } else {
+            parallelism <- "none"
+        }
+    } else {
+        parallelism <- match.arg(parallelism)
+        if (parallelism == "none" && !is.null(num_cores) && num_cores > 1) {
+            stop("contradiction between `parallelism = \"none\"` ",
+                 "and `num_cores = ", num_cores, "`; please resolve",
+                 call. = FALSE)
+        }
+        if (parallelism == "future") {
+            if (!requireNamespace("future.apply", quietly = TRUE)) {
+                stop("`parallelism = \"future\"` uses `package:future.apply`, ",
+                     "but it's not installed", call. = FALSE)
+            }
+            if (!is.null(num_cores)) {
+                stop("with `parallelism = \"future\"`, the `num_cores` ",
+                     "argument is not used to set up the backend; ",
+                     "use `future::plan` instead",
+                     call. = FALSE)
+            }
+        }
+        if (parallelism == "parallel") {
+            if (!requireNamespace("parallel", quietly = TRUE)) {
+                stop("`parallelism = \"parallel\"` uses `package:parallel`, ",
+                     "but it's not installed", call. = FALSE)
+            }
+        }
+    }
 
     ##------------------------------------
 
@@ -185,7 +223,7 @@ bootstrap_model <- function(base_model,
 
     ## deciding on random blocks. Subset of rand_cols:
     if (is.null(resample_specific_blocks)) {
-        if(length(rand_cols) > 1){
+        if (length(rand_cols) > 1) {
             entropy_levels <- unlist(lapply(rand_cols, function(rc){
                 calc_entropy(base_data[, rc])
             }))
@@ -322,13 +360,14 @@ bootstrap_model <- function(base_model,
                  orig_df = orig_df)
 }
 
+
 #' @export
 #' @rdname bootstrap_model
 #' @param suppress_loading_bar
 #' defunct now
 #' @param allow_conv_error
 #' defunct now
-BootGlmm <- function(base_model,
+BootGlmm <- function(base_model, # nocov start
                      resamples = 9999,
                      base_data = NULL,
                      return_coefs_instead = FALSE,
@@ -350,7 +389,7 @@ BootGlmm <- function(base_model,
                     narrowness_avoid = narrowness_avoid,
                     num_cores = num_cores,
                     suppress_sampling_message = suppress_sampling_message)
-}
+} # nocov end
 
 #' Runs the bootstrapping of the models
 #'
@@ -382,34 +421,21 @@ bootstrap_runner <- function(bootstrap_function,
     parallelism <- match.arg(parallelism)
 
     if (parallelism == "future") {
-        if (requireNamespace("future.apply", quietly = TRUE)) {
-            return(future.apply::future_lapply(
-                                     1:resamples,
-                                     function(i){
-                                         bootstrap_function()
-                                     }))
-        }
-        warning("future.apply not installed, using base lapply",
-                call. = FALSE)
-        parallelism <- "none"
+        ## assuming end-user sets plan elsewhere
+        return(future.apply::future_lapply(
+                                 1:resamples,
+                                 function(i){
+                                     bootstrap_function()
+                                 }))
     }
 
     if (parallelism == "parallel") {
-        if (requireNamespace("parallel", quietly = TRUE)) {
-            if (is.null(num_cores)) {
-                num_cores <- parallel::detectCores() - 1
-            }
-
-            return(parallel::mclapply(1:resamples,
-                                      function(i){
-                                          bootstrap_function()
-                                      },
-                                      mc.cores = num_cores,
-                                      mc.preschedule = FALSE))
-        }
-        warning("parallel not installed, using base lapply",
-                call. = FALSE)
-        parallelism <- "none"
+        return(parallel::mclapply(1:resamples,
+                                  function(i){
+                                      bootstrap_function()
+                                  },
+                                  mc.cores = num_cores,
+                                  mc.preschedule = FALSE))
     }
 
     lapply(1:resamples, function(i) bootstrap_function())