+++ tests/testthat/test-bootstrap_model.R
@@ -25,6 +25,10 @@ test_that("bootstrap_model fails when there is no data", {
     xy_data <- data.frame(x = x, y = y)
     simple_model <- lm(y ~ x, data = xy_data)
 
+    ## warning if there is
+    expect_warning(bootstrap_model(base_model = simple_model,
+                                   resamples = 20))
+
     simple_model_nodata <- simple_model
     simple_model_nodata$model <- NULL
 
@@ -35,7 +39,7 @@ test_that("bootstrap_model fails when there is no data", {
 })
 
 test_that("bootstrap_model works on zero-inflated models", {
-    if(!requireNamespace("glmmTMB", quietly = TRUE)){
+    if (!requireNamespace("glmmTMB", quietly = TRUE)) {
         skip("need glmmTMB to be installed for zero-inflated")
     }
 
@@ -62,4 +66,49 @@ test_that("bootstrap_model works on zero-inflated models", {
 
     expect_equal(unname(unlist(lapply(zero_boot, class))),
                  c("matrix", "matrix"))
+})
+
+test_that("bootstrap_model parallelism modes", {
+    x <- rnorm(20)
+    y <- rnorm(20)
+    xy_data <- data.frame(x = x, y = y)
+    simple_model <- lm(y ~ x, data = xy_data)
+
+    expect_error(bootstrap_model(base_model = simple_model,
+                                 base_data = xy_data,
+                                 resamples = 20,
+                                 parallelism = "none",
+                                 num_cores = 4))
+    expect_error(bootstrap_model(base_model = simple_model,
+                                 base_data = xy_data,
+                                 resamples = 20,
+                                 parallelism = "none",
+                                 num_cores = 1,
+                                 suppress_sampling_message = TRUE),
+                 NA)
+    expect_error(bootstrap_model(base_model = simple_model,
+                                 base_data = xy_data,
+                                 resamples = 20,
+                                 parallelism = "none",
+                                 num_cores = NULL,
+                                 suppress_sampling_message = TRUE),
+                 NA)
+
+    if (!requireNamespace("future.apply", quietly = TRUE)) {
+        skip("need future.apply for this next test")
+    }
+
+    expect_error(bootstrap_model(base_model = simple_model,
+                                 base_data = xy_data,
+                                 resamples = 20,
+                                 parallelism = "future",
+                                 num_cores = 4,
+                                 suppress_sampling_message = TRUE))
+    expect_error(bootstrap_model(base_model = simple_model,
+                                 base_data = xy_data,
+                                 resamples = 20,
+                                 parallelism = "future",
+                                 num_cores = NULL,
+                                 suppress_sampling_message = TRUE),
+                 NA)
 })